Refactor schemas, move base and core classes to a new module
The core and dataset related base classes, primarily 'StrictCamelBaseModel' and 'BaseDataset', have been moved to a new module named base.py. The imports of these classes in various schema files have been updated accordingly. This refactor simplifies the structure and improves the organization of schema related classes.
