+++ src/mdio/schemas/v1/dataset.py
@@ -3,7 +3,7 @@
 
 from pydantic import Field
 
-from mdio.schemas.dataset import BaseDataset
+from mdio.schemas.base import BaseDataset
 from mdio.schemas.dimension import NamedDimension
 from mdio.schemas.metadata import UserAttributes
 from mdio.schemas.v1.variable import Variable
