+++ src/mdio/schemas/v1/stats.py
@@ -20,7 +20,7 @@ from typing import TypeAlias
 
 from pydantic import Field
 
-from mdio.schemas.core import StrictCamelBaseModel
+from mdio.schemas.base import StrictCamelBaseModel
 from mdio.schemas.metadata import VersionedMetadataConvention
